# server/swagger.json
{
  "basePath": "/api",
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "version": "v1.0",
    "title": "Items Server"
  },
  "securityDefinitions": {
    "Token": {
      "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/hotels": {
      "get": {
        "summary": "Show All Items",
        "description": "Show All Items",
        "tags": [
          "hotels"
        ],
        "operationId": "ShowAllItems",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          },
          "500": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "summary": "Create an Hotel",
        "description": "Create an Hotel",
        "tags": [
          "hotels"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "hotel",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo ok"
          },
          "500": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/hotels/{id}": {
      "delete": {
        "summary": "",
        "operationId": "getHotelById",
        "tags": [
          "hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the hotel",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo ok",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          },
          "500": {
            "description": "Unexpected error"
          }
        }
      },
      "put": {
        "summary": "",
        "operationId": "",
        "tags": [
          "hotels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "hotel",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo ok",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          },
          "500": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/send_mail": {
      "post": {
        "summary": "Send Mail",
        "tags": [
          "Mailer"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "email",
            "schema": {
              "$ref": "#/definitions/hotel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail enviado correctamente"
          },
          "401": {
            "description": "Error al enviar el mail"
          }
        }
      }
    }
  },
  "definitions": {
    "hotel": {
      "type": "object",
      "properties": {
        "slug": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "img": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "mail": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string",
          "default": "juanluis.belda@gmail.com"
        },
        "to": {
          "type": "string"
        },
        "subject": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    }
  }
}